// src/types/style.types.ts
export type Flex = {
  justify?: string;
  align?: string;
  row?: string;
  height?: string;
  margin?: string;
  width?: string;
  position?: string;
};

export type Div = {
  height?: string;
  margin?: string;
  width?: string;
  padding?: string;
  size?: string;
  bw?: string;
  bs?: string;
  color?: string;
};

export type MainDiv = {
  padding?: string;
};

export type Span = {
  margin?: string;
  padding?: string;
  size?: string;
  color?: string;
  weight?: string;
  backColor?: string;
};